integrated versioning into service APIs
I started with the following strategy using Versionist:

- APIs are versioned
- version can be specified via request param (so can be tested in the browser etc.)
- A specific version can be default if no version # specified
-- Only V1 right now

on URI versioning :) - after reading  a bunch more about it and thinking about it, I decided to keep a nice, clean resource URI with a default version and an optional request parameter which will probably correspond to 'current' or 'stable' version.  Actually the ruby gem I used (versionist) lets me incorporate > 1 strategy (any combination of HTTP header, Path or Request Parameter) if I change my mind.

So in my case, this works

# get hotel with ID 1
http://lvh.me:3000/hotels/1?version=1

and this works (same result as V1 is the default route)

http://lvh.me:3000/hotels/1/

and this works (get all hotels)

http://lvh.me:3000/hotels/

as well as this

http://lvh.me:3000/hotels?version=1

once we have other versions, they will work (and can be the default version) as well